4th commit

+++ ft_strlen.c
@@ -1,6 +1,8 @@
-int		ft_strlen(const char *str)
+#include <string.h>
+
+size_t	ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str && str[i])
