+++ ft_strncmp.c
@@ -1,6 +1,8 @@
-int		ft_strncmp(const char *s1, const char *s2, int n)
+#include <string.h>
+
+int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (n == 0)
